Add UDP + DHCP + DNS (Phases 50-51), Phase 60 syscalls, PS/2 sign fix
Networking:
- kernel/net/udp.c — arch-agnostic UDP sockets in the tcp.c mold:
  8 sockets, per-socket rx queue, ephemeral ports, udp_bind/sendto/
  recvfrom/close. IP glue in both kernels (proto 17 dispatch from the
  timer-IRQ rx path; transmit allows src 0.0.0.0 and 255.255.255.255
  broadcast for DHCP).
- DHCP client: full DISCOVER/OFFER/REQUEST/ACK with xid matching,
  option parsing (mask/router/dns) and 3 retries. The `dhcp` CLI
  command applies the lease via kernel_net_apply_config(), which
  updates net_ip/net_gw and re-registers TCP/UDP local addresses.
- DNS client: A-record queries with compression-pointer handling;
  server defaults to slirp's 10.0.2.3, overridden by DHCP. `nslookup`
  CLI command.
- New CLI: udpbind, udpsend, udprecv, udpclose, dhcp, nslookup.
- Fixed cli_printf: no %u support — silently dropped the specifier
  (tcpstat's IP columns were blank because of this too).
- Makefile: udp.o wired into both kernels (the x64 bin rule's explicit
  dependency list needed the new object too, not just the link list).

Phase 60 syscalls (both arches): SYS_SETUID/SETGID/GETUID/GETGID/
CHDIR/GETCWD/UMASK (79-85). Per-task uid/gid/umask/cwd fields,
inherited across fork; setuid/setgid gated on root; chdir validates
the path against the VFS; getcwd copies into a validated user buffer.
userlib wrappers added. getty/login don't call these yet.

PS/2 mouse: movement is 9-bit two's complement with sign bits in the
flags byte (bits 4/5) — the handler previously truncated to int8, so
any single delta beyond +-127 corrupted. FFI widened to i32 through
mouse.rs/gui.

Verified (QEMU slirp, both arches where applicable): dhcp lease
10.0.2.15/24 gw 10.0.2.2 dns 10.0.2.3; nslookup example.com resolves
via real host DNS; UDP datagram received by a host listener; ping
works after reconfiguration; TCP suite 11/11; GUI smoke test clean.

+++ kernel/core/verniskernel/src/gui/mod.rs
@@ -307,7 +307,7 @@ pub unsafe extern "C" fn gui_handle_key(scancode: u8, _pressed: u8) {
 /// This processes the raw PS/2 packet, updates mouse position,
 /// and generates high-level events.
 #[no_mangle]
-pub unsafe extern "C" fn gui_handle_mouse(flags: u8, dx: i8, dy: i8) {
+pub unsafe extern "C" fn gui_handle_mouse(flags: u8, dx: i32, dy: i32) {
     // Update mouse position via the mouse module
     mouse::mouse_handle_packet(flags, dx, dy);
